[WIP] Support tensors as Dict keys
This prepares the PR where we implement sets in terms of dicts.

The Hashable variable is opaque on purpose so that it fails hard if
if it unadvertedly leaks back into user code.

We also found and fixed a number of latent bugs and inconsistencies
in the way dynamo checked what can be a dict key.

Fixes https://github.com/pytorch/pytorch/issues/111603

ghstack-source-id: 655dc228396b8402becfad00ec07ae8810e556fa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/111196