[pytree] Add optimized `tree_leaves` implementation
pytree is used in many hot paths for dynamo tracing and in many cases we don't
care about the tree spec and just want the flattened list. This improves
`pytree.tree_leaves` to not construct the spec which gives a noticeable
performance improvement when multiplies by the millions of times it gets called
during tracing.

ghstack-source-id: 475292a240bf48be9dacf19f97c58cc2d8a9fb55
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112323